refactor: extract non-lnd specific util methods
