+++ app/lib/lnd/lightning.js
@@ -5,7 +5,8 @@ import { loadSync } from '@grpc/proto-loader'
 import { BrowserWindow } from 'electron'
 import StateMachine from 'javascript-state-machine'
 import LndConfig from './config'
-import { getDeadline, validateHost, createSslCreds, createMacaroonCreds, waitForFile } from './util'
+import { getDeadline, createSslCreds, createMacaroonCreds, waitForFile } from './util'
+import { validateHost } from '../utils/validateHost'
 import methods from './methods'
 import { mainLog } from '../utils/log'
 import subscribeToTransactions from './subscribe/transactions'
