+++ app/lib/lnd/walletUnlocker.js
@@ -4,7 +4,8 @@ import grpc from 'grpc'
 import { loadSync } from '@grpc/proto-loader'
 import StateMachine from 'javascript-state-machine'
 import LndConfig from './config'
-import { getDeadline, validateHost, createSslCreds } from './util'
+import { getDeadline, createSslCreds } from './util'
+import { validateHost } from '../utils/validateHost'
 import methods from './walletUnlockerMethods'
 import { mainLog } from '../utils/log'
